requirements.txt
fastapi
pydantic
uvicorn
python-dotenv
langchain-openai
langchain-community
langchain-text-splitters
faiss-cpu
sentence-transformers
